pythonStompWorker/requirements.txt
attrs==17.3.0
docopt==0.6.2
pkg-resources==0.0.0
pluggy==0.6.0
py==1.5.2
pytest==3.3.0
six==1.11.0
stomp.py==4.1.19
